Pair programming with Felipe to coordiate all changes. Made connection between Motorcycle model and OrderItem model. Modified routes with the new path of the order controller. Moved order controller file into user folder in Controllers. Moved order views files into user folder in views.
